# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital API",
        "description": "API documentation for the Hospital application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/summary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get system summary statistics",
                "operationId": "e04ac99c958d75bb1acbe01eeef47c67",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "doctors": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "patients": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "reviews": {
                                            "type": "integer",
                                            "example": 1200
                                        },
                                        "revenue": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 125000.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/analytics": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get filtered analytics data",
                "operationId": "e1ceeb413a44088eae3d681ca0e5639a",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of analytics (payments|reviews|registrations)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "payments",
                                "reviews",
                                "registrations"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "count": {
                                                "type": "integer"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/doctor-stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get top performing doctors",
                "operationId": "2cc2d71019f46b8d803df10dcc16e9aa",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of doctors to return",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Doctor"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/diagnostics": {
            "get": {
                "tags": [
                    "Diagnostics"
                ],
                "summary": "List diagnostics",
                "operationId": "a9f0810630513f1a6034a6891b605d60",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Diagnostic"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Diagnostics"
                ],
                "summary": "Create diagnostic",
                "operationId": "6ffe56d6edaf454e724f29204acc19ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDiagnosticRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diagnostic"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/diagnostics/{id}": {
            "get": {
                "tags": [
                    "Diagnostics"
                ],
                "summary": "Get diagnostic by ID",
                "operationId": "8a2637e1d0a71a68c53db7551a214e1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diagnostic"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Diagnostics"
                ],
                "summary": "Update diagnostic",
                "operationId": "d0a09e99570f36b452b52f8886f2ce72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDiagnosticRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diagnostic"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Diagnostics"
                ],
                "summary": "Delete diagnostic",
                "operationId": "1df6f3b1c5757af753150360f2ddbabe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreDiagnosticRequest": {
                "properties": {
                    "prescription": {
                        "type": "string"
                    },
                    "report_id": {
                        "type": "integer"
                    },
                    "diagnostics": {
                        "type": "string"
                    },
                    "patient_id": {
                        "type": "integer"
                    },
                    "doctor_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateDiagnosticRequest": {
                "properties": {
                    "prescription": {
                        "type": "string"
                    },
                    "report_id": {
                        "type": "integer"
                    },
                    "diagnostics": {
                        "type": "string"
                    },
                    "patient_id": {
                        "type": "integer"
                    },
                    "doctor_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Diagnostic": {
                "title": "Diagnostic",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "prescription": {
                        "type": "string",
                        "example": "Prescription"
                    },
                    "report_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "diagnostics": {
                        "type": "string",
                        "example": "Diagnostics"
                    },
                    "patient_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "doctor_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Doctor": {
                "title": "Doctor",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "specialization": {
                        "type": "string",
                        "example": "Cardiology"
                    },
                    "experience": {
                        "type": "string",
                        "example": "10 years"
                    },
                    "education": {
                        "type": "string",
                        "example": "MD, PhD"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male"
                    },
                    "city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "state": {
                        "type": "string",
                        "example": "New York"
                    },
                    "country": {
                        "type": "string",
                        "example": "USA"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "bio": {
                        "type": "string",
                        "example": "This is a great doctor!"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "department": {
                        "type": "string",
                        "example": "Cardiology"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}